fix: fixed buttons and change the project structure
