+++ Client/src/index.css
@@ -24,7 +24,7 @@ html {
   background: linear-gradient(120deg,
       rgba(0, 194, 203, 1) 0%,
       rgba(2, 88, 82, 1) 100%);
-  padding: 10px 20px;
+  padding: 7px 15px;
   border-radius: 30px;
   color: rgba(255, 255, 255, 1);
   font-weight: 600;
